+++ src/main/java/com/pitchain/common/application/EntityFacade.java
@@ -11,8 +11,8 @@ import com.pitchain.common.security.MemberDetails;
 import com.pitchain.company.domain.Company;
 import com.pitchain.individual.domain.Individual;
 import com.pitchain.member.domain.Member;
-import com.pitchain.company.application.CompanyRepository;
-import com.pitchain.individual.application.IndividualRepository;
+import com.pitchain.company.infrastructure.CompanyRepository;
+import com.pitchain.individual.infrastructure.IndividualRepository;
 import com.pitchain.member.infrastucture.MemberRepository;
 import com.pitchain.sp.domain.Sp;
 import com.pitchain.sp.infrastucture.SpRepository;
