+++ src/test/java/com/pitchain/service/MemberServiceTest.java
@@ -12,8 +12,8 @@ import com.pitchain.individual.domain.Individual;
 import com.pitchain.member.domain.Member;
 import com.pitchain.common.security.MemberDetails;
 import com.pitchain.member.application.MemberService;
-import com.pitchain.company.application.CompanyRepository;
-import com.pitchain.individual.application.IndividualRepository;
+import com.pitchain.company.infrastructure.CompanyRepository;
+import com.pitchain.individual.infrastructure.IndividualRepository;
 import com.pitchain.member.infrastucture.MemberRepository;
 import com.pitchain.upload.application.S3Service;
 import com.pitchain.util.EntitySaver;
