+++ src/test/java/com/pitchain/util/EntitySaver.java
@@ -10,8 +10,8 @@ import com.pitchain.investment.infrastucture.InvestmentRepository;
 import com.pitchain.company.domain.Company;
 import com.pitchain.individual.domain.Individual;
 import com.pitchain.member.domain.Member;
-import com.pitchain.company.application.CompanyRepository;
-import com.pitchain.individual.application.IndividualRepository;
+import com.pitchain.company.infrastructure.CompanyRepository;
+import com.pitchain.individual.infrastructure.IndividualRepository;
 import com.pitchain.member.infrastucture.MemberRepository;
 import com.pitchain.mybmhistory.infrastructure.MyBmHistoryRepository;
 import com.pitchain.mysphistory.infrastucture.MySpHistoryRepository;
